move integration-like tests from comment.rs to own file
